add button navbar
