+++ react/src/pages/single-board/nav-bars/sidebar.scss
@@ -24,3 +24,17 @@
     }
   }
   
+  .cardMoreNav {
+    position: absolute;
+    z-index: 1000;
+    margin: 10px;
+    background-color: white;
+    border-radius: 5px;
+    height: 40px;
+    width: 130px;
+    box-shadow: 0px 0px 5px rgba(75, 75, 75, 0.2);
+  }
+
+  .cardMoreButton {
+    margin: 5px;
+  }
