+++ react/src/pages/single-board/nav-bars/sidebar.tsx
@@ -1,8 +1,8 @@
 
 import "./sidebar.scss"
 
-import { ButtonArrow } from "../../buttons/buttonArrow"
-import { ButtonCard } from "../../buttons/buttonCard"
+import { ButtonArrow } from "./buttons/buttonArrow"
+import { ButtonCard } from "./buttons/buttonCard"
 
 export const Sidebar = () => {
